Move PsiUtil to pakage utils
