+++ src/main/java/org/ml_methods_group/utils/PSIUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ml_methods_group.algorithm;
+package org.ml_methods_group.utils;
 
 import com.intellij.psi.PsiClass;
 import com.intellij.psi.PsiMethod;
